Changes in body
Who we are?
Maverick Solution is revolution in Digital Consulting and Human Resource. We provide valid job search assistance to all over India, soon to the World.

What we do?
As of now Maverick Solution is running on private funds, this the sole reason we take consultancy and service charges from candidate. But don't worry if you are not selected for a particular role you don't need to pay anything at all.
#We seek candidate who are eligible for certain job post.
#Provide them necessary aid to prepare for the interview.
#Live screening of the candidate to brush-up their skills.
#Prepare and select candidate for right job profile.

What are our goals?
Our only goal to be successful by making you successful. We want your opinion to be our good will. Feel free to write us for any feedback at : info@maverickhr.in